Bookmarks now fetch and store page titles.
Use a before_create hook to grab the title such that it can be stored.

This perhaps should be done asynchronously, maybe with sidekiq or resque,
if this type of thing is done on a large scale web app. For the purposes
of this app, the blocking request should be fine.